Changed vagrant file to add a custom domain name. took out the node name in the site_xe11g.pp and added swap to fstab
Made three changes to the 2 files. Added the custom domain name.  This
is done by vagrant by default.  However, we were setting it with the
node in the class in the site_xe11g.pp So I took that node name out.
That is for puppet client server. I have found out how to add swap to
fstab so that when we halt...the swap doesn't get lost...yeah!